LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-rtl.tentative-expected.txt


FAIL scroll-initial-target reflects vertical rtl writing mode. assert_equals: scroller is vertically scrolled to scroll-initial-target expected 291 but got 0
